refactor(ui): extract CommandBarInChat and extend plan a11y
Move the file-change command bar out of SidebarChat into composer/CommandBarInChat
and add aria labels for accept/reject controls, file links, and plan expand/collapse.
